move add_preview_of_content_to_notifications and get_preview_of_content to dashboard.py
